Move copy action into separate helper file
