+++ frontend/app_v2/src/components/Search/SearchData.js
@@ -4,6 +4,7 @@ import { useLocation } from 'react-router-dom'
 
 import useGetSite from 'common/useGetSite'
 import searchApi from 'services/api/search'
+import { copy } from 'common/actionHelpers'
 
 /**
  * @summary SearchData
@@ -67,35 +68,6 @@ function SearchData() {
     setCurrentFilter(filter)
   }
 
-  const actions = [
-    {
-      actionTitle: 'copy',
-      iconName: 'Copy',
-      confirmationMessage: 'Copied!',
-      clickHandler: function clickCopyHandler(str) {
-        // Create new element
-        const el = document.createElement('textarea')
-        // Set value (string to be copied)
-        el.value = str
-        // Set non-editable to avoid focus and move outside of view
-        el.setAttribute('readonly', '')
-        el.style = { position: 'absolute', left: '-9999px' }
-        document.body.appendChild(el)
-        // Select text inside element
-        el.select()
-        // Copy text to clipboard
-        document.execCommand('copy')
-        // Remove temporary element
-        document.body.removeChild(el)
-        //Set tooltip confirmation with timeout
-        document.getElementById(`copy-message-${str}`).style.display = 'contents'
-        setTimeout(function timeout() {
-          document.getElementById(`copy-message-${str}`).style.display = 'none'
-        }, 1000)
-      },
-    },
-  ]
-
   return {
     currentFilter,
     siteTitle: title ? title : 'FirstVoices',
@@ -104,7 +76,7 @@ function SearchData() {
     handleFilter,
     isLoading,
     items,
-    actions,
+    actions: [copy],
     searchTerm,
   }
 }
